[sfmData] Rename `exposureSetting.h` to `ExposureSetting.h`
