+++ src/aliceVision/sfmData/ExposureSetting.hpp
